+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/api/usecase/pages/catalog/UseCaseCatalogApiService.java
@@ -1,8 +1,22 @@
+/*
+ *  Copyright (c) 2024 sovity GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       sovity GmbH - initial API and implementation
+ *
+ */
+
 package de.sovity.edc.ext.wrapper.api.usecase.pages.catalog;
 
 import de.sovity.edc.ext.wrapper.api.ui.model.UiDataOffer;
 import de.sovity.edc.ext.wrapper.api.ui.pages.catalog.UiDataOfferBuilder;
-import de.sovity.edc.ext.wrapper.api.usecase.model.CatalogQueryParams;
+import de.sovity.edc.ext.wrapper.api.usecase.model.CatalogQuery;
 import de.sovity.edc.utils.catalog.DspCatalogService;
 import lombok.RequiredArgsConstructor;
 import org.eclipse.edc.spi.query.QuerySpec;
@@ -15,14 +29,14 @@ public class UseCaseCatalogApiService {
     private final DspCatalogService dspCatalogService;
     private final FilterExpressionMapper filterExpressionMapper;
 
-    public List<UiDataOffer> fetchDataOffers(CatalogQueryParams catalogQueryParams) {
-        var querySpec = buildQuerySpec(catalogQueryParams);
+    public List<UiDataOffer> fetchDataOffers(CatalogQuery catalogQuery) {
+        var querySpec = buildQuerySpec(catalogQuery);
 
-        var dspCatalog = dspCatalogService.fetchDataOffersWithFilters(catalogQueryParams.getTargetEdc(), querySpec);
+        var dspCatalog = dspCatalogService.fetchDataOffersWithFilters(catalogQuery.getTargetEdc(), querySpec);
         return uiDataOfferBuilder.buildUiDataOffers(dspCatalog);
     }
 
-    private QuerySpec buildQuerySpec(CatalogQueryParams params) {
+    private QuerySpec buildQuerySpec(CatalogQuery params) {
         var builder = QuerySpec.Builder.newInstance()
                 .limit(withDefault(params.getLimit(), Integer.MAX_VALUE))
                 .offset(withDefault(params.getOffset(), 0));